# Remove the use of `testUtil` from within `drake/systems`