--- a/drake/systems/plants/test/bullet_collision_zero_rad_sphere_test.cpp
+++ b/drake/systems/plants/test/bullet_collision_zero_rad_sphere_test.cpp
@@ -3,7 +3,6 @@
 #include "drake/common/drake_path.h"
 #include "drake/systems/plants/parser_urdf.h"
 #include "drake/systems/plants/RigidBodyTree.h"
-#include "drake/util/testUtil.h"
 
 using Eigen::Matrix3Xd;
 using Eigen::VectorXd;