--- a/drake/systems/plants/test/compareParsersmex.cpp
+++ b/drake/systems/plants/test/compareParsersmex.cpp
@@ -7,7 +7,6 @@
 #include "drake/common/eigen_types.h"
 #include "drake/systems/plants/RigidBodyTree.h"
 #include "drake/util/drakeMexUtil.h"
-#include "drake/util/testUtil.h"
 
 using namespace Eigen;
 using namespace std;