--- a/drake/systems/plants/test/compareRigidBodySystems.cpp
+++ b/drake/systems/plants/test/compareRigidBodySystems.cpp
@@ -6,7 +6,6 @@
 
 #include "drake/common/eigen_matrix_compare.h"
 #include "drake/systems/plants/RigidBodyFrame.h"
-#include "drake/util/testUtil.h"
 
 using std::make_shared;
 using drake::RigidBodySystem;