--- a/drake/systems/plants/test/lidarTest.cpp
+++ b/drake/systems/plants/test/lidarTest.cpp
@@ -1,10 +1,11 @@
 #include <cmath>
 
+#include "gtest/gtest.h"
+
 #include "drake/common/drake_path.h"
 #include "drake/systems/LCMSystem.h"
 #include "drake/systems/plants/BotVisualizer.h"
 #include "drake/systems/plants/RigidBodySystem.h"
-#include "drake/util/testUtil.h"
 #include "lcmtypes/bot_core/planar_lidar_t.hpp"
 
 using Eigen::VectorXd;
@@ -13,7 +14,7 @@ using std::make_shared;
 namespace drake {
 namespace {
 
-int do_main(int argc, char* argv[]) {
+GTEST_TEST(LidarTest, BasicTest) {
   auto rigid_body_sys = make_shared<RigidBodySystem>();
   rigid_body_sys->AddModelInstanceFromFile(
       GetDrakePath() + "/systems/plants/test/lidarTest.sdf", DrakeJoint::FIXED);
@@ -49,39 +50,34 @@ int do_main(int argc, char* argv[]) {
       "ERROR: LIDAR scanner says it is vertical but is not!");
   }
 
-  valuecheck(static_cast<size_t>(1), lidar_sensor->num_pixel_rows(),
-             std::string("Problems with num_pixel_rows(). "));
-  valuecheck(static_cast<size_t>(640), lidar_sensor->num_pixel_cols(),
-             std::string("Problems with num_pixel_cols(). "));
+  EXPECT_EQ(static_cast<size_t>(1), lidar_sensor->num_pixel_rows());
+  EXPECT_EQ(static_cast<size_t>(640), lidar_sensor->num_pixel_cols());
 
-  valuecheck(0, lidar_sensor->min_pitch(),
+  EXPECT_NEAR(0, lidar_sensor->min_pitch(),
              std::numeric_limits<double>::epsilon());
 
-  valuecheck(0, lidar_sensor->max_pitch(),
+  EXPECT_NEAR(0, lidar_sensor->max_pitch(),
              std::numeric_limits<double>::epsilon());
 
-  valuecheck(-1.7, lidar_sensor->min_yaw(),
+  EXPECT_NEAR(-1.7, lidar_sensor->min_yaw(),
              std::numeric_limits<double>::epsilon());
 
-  valuecheck(1.7, lidar_sensor->max_yaw(),
+  EXPECT_NEAR(1.7, lidar_sensor->max_yaw(),
              std::numeric_limits<double>::epsilon());
 
-  valuecheck(0.08, lidar_sensor->min_range(),
+  EXPECT_NEAR(0.08, lidar_sensor->min_range(),
              std::numeric_limits<double>::epsilon());
 
-  valuecheck(25, lidar_sensor->max_range(),
+  EXPECT_NEAR(25, lidar_sensor->max_range(),
              std::numeric_limits<double>::epsilon());
 
   double t = 0;
   VectorXd x = VectorXd::Zero(rigid_body_sys->getNumStates());
   VectorXd u = VectorXd::Zero(rigid_body_sys->getNumInputs());
 
-  //  rigid_body_sys->getRigidBodyTree()->drawKinematicTree("/tmp/lidar.dot");
-
   auto distances = rigid_body_sys->output(t, x, u);
 
-  valuecheck(640, static_cast<int>(distances.size()),
-             std::string("Problem with number of distances. "));
+  EXPECT_EQ(640, static_cast<int>(distances.size()));
 
   const double min_yaw = lidar_sensor->min_yaw();
   const double max_yaw = lidar_sensor->max_yaw();
@@ -95,13 +91,13 @@ int do_main(int argc, char* argv[]) {
     // computation of the analytical distances.
     if (std::abs(theta) >=
         M_PI / 2.0 + tol) {  // Should not be hitting any wall.
-      valuecheck(max_range, distances(i));
+      EXPECT_EQ(max_range, distances(i));
     } else if (theta <= -M_PI / 4) {  // hitting the right wall
-      valuecheck(-1.0 / std::sin(theta), distances(i), tol);
+      EXPECT_NEAR(-1.0 / std::sin(theta), distances(i), tol);
     } else if (theta >= M_PI / 4) {  // hitting the left wall
-      valuecheck(1.0 / std::sin(theta), distances(i), tol);
+      EXPECT_NEAR(1.0 / std::sin(theta), distances(i), tol);
     } else {  // hitting the front wall
-      valuecheck(1 / std::cos(theta), distances(i), tol);
+      EXPECT_NEAR(1 / std::cos(theta), distances(i), tol);
     }
   }
 
@@ -124,16 +120,7 @@ int do_main(int argc, char* argv[]) {
   for (int i = 0; i < distances.size(); i++) msg.ranges.push_back(distances(i));
 
   lcm->publish("DRAKE_PLANAR_LIDAR_0_rear_wall", &msg);
-
-  return 0;
 }
 
 }  // namespace
 }  // namespace drake
-
-/**
- * This unit test sets up a LIDAR in a box room and verifies the returns.
- */
-int main(int argc, char* argv[]) {
-  return drake::do_main(argc, argv);
-}