--- a/drake/systems/plants/test/testMagnetometer.cpp
+++ b/drake/systems/plants/test/testMagnetometer.cpp
@@ -4,7 +4,6 @@
 #include "drake/common/eigen_matrix_compare.h"
 #include "drake/math/roll_pitch_yaw.h"
 #include "drake/systems/plants/RigidBodySystem.h"
-#include "drake/util/testUtil.h"
 
 using Eigen::Vector3d;
 using Eigen::Vector4d;