--- a/drake/systems/plants/test/testMassSpringDamper.cpp
+++ b/drake/systems/plants/test/testMassSpringDamper.cpp
@@ -3,7 +3,6 @@
 #include "drake/common/drake_path.h"
 #include "drake/common/eigen_matrix_compare.h"
 #include "drake/systems/plants/RigidBodySystem.h"
-#include "drake/util/testUtil.h"
 
 using Eigen::Matrix;
 using std::make_shared;