--- a/drake/systems/test/vector_test.cc
+++ b/drake/systems/test/vector_test.cc
@@ -4,7 +4,6 @@
 
 #include "drake/common/eigen_matrix_compare.h"
 #include "drake/systems/test/pendulum.h"
-#include "drake/util/testUtil.h"
 
 using drake::CombinedVector;
 using drake::size;