--- a/drake/systems/trajectories/test/CMakeLists.txt
+++ b/drake/systems/trajectories/test/CMakeLists.txt
@@ -3,7 +3,7 @@ target_link_libraries(testPiecewisePolynomial drakeTrajectories ${GTEST_BOTH_LIB
 add_test(NAME testPiecewisePolynomial COMMAND testPiecewisePolynomial)
 
 add_executable(testExponentialPlusPiecewisePolynomial testExponentialPlusPiecewisePolynomial.cpp)
-target_link_libraries(testExponentialPlusPiecewisePolynomial drakeTrajectories)
+target_link_libraries(testExponentialPlusPiecewisePolynomial drakeTrajectories ${GTEST_BOTH_LIBRARIES})
 add_test(NAME testExponentialPlusPiecewisePolynomial COMMAND testExponentialPlusPiecewisePolynomial)
 
 add_matlab_test(NAME systems/trajectories/test/expplusppFnderTest COMMAND expplusppFnderTest)