--- a/drake/systems/trajectories/test/testExponentialPlusPiecewisePolynomial.cpp
+++ b/drake/systems/trajectories/test/testExponentialPlusPiecewisePolynomial.cpp
@@ -1,14 +1,16 @@
 #include "drake/systems/trajectories/ExponentialPlusPiecewisePolynomial.h"
-#include <Eigen/Core>
-#include <random>
-#include <iostream>
+
 #include <cmath>
-#include "drake/util/testUtil.h"
+#include <random>
+
+#include <Eigen/Core>
+#include "gtest/gtest.h"
 
 using namespace std;
 using namespace Eigen;
 
-default_random_engine generator;
+namespace drake {
+namespace {
 
 template <typename CoefficientType>
 void testSimpleCase() {
@@ -23,6 +25,7 @@ void testSimpleCase() {
   MatrixX A = MatrixX::Random(1, 1);
   MatrixX alpha = MatrixX::Random(1, 1);
 
+  default_random_engine generator;
   auto segment_times =
       PiecewiseFunction::randomSegmentTimes(num_segments, generator);
   auto polynomial_part = PiecewisePolynomial<CoefficientType>::random(
@@ -43,12 +46,13 @@ void testSimpleCase() {
       K(0) * A(0) * std::exp(A(0) * (t - expPlusPp.getStartTime())) * alpha(0) +
       polynomial_part.derivative().scalarValue(t);
 
-  valuecheck(check, expPlusPp.value(t)(0), 1e-8);
-  valuecheck(derivative_check, derivative.value(t)(0), 1e-8);
+  EXPECT_NEAR(check, expPlusPp.value(t)(0), 1e-8);
+  EXPECT_NEAR(derivative_check, derivative.value(t)(0), 1e-8);
 }
 
-int main(int argc, char **argv) {
+GTEST_TEST(testExponentialPlusPiecewisePolynomial, BasicTest) {
   testSimpleCase<double>();
-  std::cout << "test passed";
-  return 0;
 }
+
+}  // namespace
+}  // namespace drake