--- a/drake/util/testUtil.h
+++ b/drake/util/testUtil.h
@@ -1,29 +1,11 @@
 #pragma once
 
-#include <chrono>
 #include <cmath>
 #include <sstream>
 #include <stdexcept>
 #include <string>
 
 #include <Eigen/Core>
-
-// requires <chrono>, which isn't available in MSVC2010...
-template <typename TimeT = std::chrono::milliseconds>
-struct measure {
-  template <typename F, typename... Args>
-  static typename TimeT::rep execution(F func, Args&&... args) {
-    auto start = std::chrono::system_clock::now();
-
-    // Now call the function with all the parameters you need.
-    func(std::forward<Args>(args)...);
-
-    auto duration = std::chrono::duration_cast<TimeT>(
-        std::chrono::system_clock::now() - start);
-
-    return duration.count();
-  }
-};
 
 template <typename Derived>
 std::string to_string(const Eigen::MatrixBase<Derived>& a) {