… drake/systems/plants/test/CMakeLists.txt