@liangfok 2016-08-27
Overall, this PR does a wonderful and significant cleanup.

However, it also switches from measuring execution time in milliseconds to seconds. Has this change impacted measurement precision? It might if the user is saving the measured values in text log files since typically only a limited number of decimal places are saved. Thoughts?

Why not use `std::high_resolution_clock()`?

---

Reviewed 20 of 20 files at r1.
Review status: all files reviewed at latest revision, all discussions resolved.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/robotlocomotion/drake/3276#-:-KQ8hIhXv2e27bXTWHaM:bm5qf6)*
<!-- Sent from Reviewable.io -->