@jwnimmer-tri 2016-08-27
> [This] switches from measuring execution time in milliseconds to seconds.

The units of reporting are changed, but the measurements are still done with sub-second precision.  I ran the RBT benchmark before and after this edit, and the results were similar:

Before:
```
no gradients: 3.336 ms
autodiff fixed max size: 28.663 ms
...
```

After:
```
no gradients: 0.00330296 s
autodiff fixed max size: 0.0286834 s
...
```

> Has this change impacted measurement precision? It might if the user is saving the measured values in text log files since typically only a limited number of decimal places are saved. Thoughts?

Given the above sample output, I don't think this is a concern.

> Why not use std::high_resolution_clock()?

When I quickly glanced at it, that block did not appear to be guaranteed to be monotonic.

---

Review status: all files reviewed at latest revision, all discussions resolved.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/robotlocomotion/drake/3276#-:-KQ8kAX-Ii8sz11rLbtY:b-kpwtrr)*
<!-- Sent from Reviewable.io -->