@liangfok 2016-08-27
In the spirit and recognition that we are in a sprint, I bestow upon this PR a <img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/> under one condition:

1. Add a CHANGELOG.md entry indicating the API change for measuring method execution times and the fact that the measurement unit was changed from milliseconds to seconds.

---

Review status: all files reviewed at latest revision, all discussions resolved.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/robotlocomotion/drake/3276#-:-KQ8pYHS62pHkJ9FKa5k:bxrcu0r)*
<!-- Sent from Reviewable.io -->