@jwnimmer-tri 2016-08-28
Done.

---

Review status: 20 of 21 files reviewed at latest revision, all discussions resolved.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/robotlocomotion/drake/3276#-:-KQDU1jX7xVlHzZ_P1dR:b-896fix)*
<!-- Sent from Reviewable.io -->