@liangfok 2016-08-28
Reviewed 1 of 1 files at r2.
Review status: all files reviewed at latest revision, 1 unresolved discussion.

---

*[CHANGELOG.md, line 26 \[r2\]](https://reviewable.io:443/reviews/robotlocomotion/drake/3276#-KQDYg0Whpnp1tRJs1-g:-KQDYg0Whpnp1tRJs1-h:b2cxwkv) ([raw file](https://github.com/robotlocomotion/drake/blob/ffdefa7382fe2545daf5c2f05f725155133760e2/CHANGELOG.md#L26)):*
> ```Markdown
> ### Changed
> 
>  - [#3276][] The measure::execution function has been replaced by MeasureExecutionTime.
> ```

BTW: nit: Surround C++ class / method names with back ticks so they appear in a different font, which indicate they are actual software artifacts.

Also add a sentence saying the units changed from milliseconds to seconds.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/robotlocomotion/drake/3276)*
<!-- Sent from Reviewable.io -->