@jwnimmer-tri 2016-08-28
Review status: all files reviewed at latest revision, 1 unresolved discussion.

---

*[CHANGELOG.md, line 26 \[r2\]](https://reviewable.io:443/reviews/robotlocomotion/drake/3276#-KQDYg0Whpnp1tRJs1-g:-KQD_APvGwicpBF1tZj-:b-p685up) ([raw file](https://github.com/robotlocomotion/drake/blob/ffdefa7382fe2545daf5c2f05f725155133760e2/CHANGELOG.md#L26)):*
<details><summary><i>Previously, liangfok (Chien-Liang Fok) wrote…</i></summary>
> BTW: nit: Surround C++ class / method names with back ticks so they appear in a different font, which indicate they are actual software artifacts.
> 
> Also add a sentence saying the units changed from milliseconds to seconds.
</details>
The prior units were undocumented; the new units are documented; the changelog should be as brief as reasonably possible; I don't think the units detail is worth writing about here. 

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/robotlocomotion/drake/3276)*
<!-- Sent from Reviewable.io -->