@liangfok 2016-08-28
Review status: all files reviewed at latest revision, 1 unresolved discussion.

---

*[CHANGELOG.md, line 26 \[r2\]](https://reviewable.io:443/reviews/robotlocomotion/drake/3276#-KQDYg0Whpnp1tRJs1-g:-KQD_obUg1WnkTZk9rDY:b-jufc4y) ([raw file](https://github.com/robotlocomotion/drake/blob/ffdefa7382fe2545daf5c2f05f725155133760e2/CHANGELOG.md#L26)):*
<details><summary><i>Previously, jwnimmer-tri (Jeremy Nimmer) wrote…</i></summary>
> The prior units were undocumented; the new units are documented; the changelog should be as brief as reasonably possible; I don't think the units detail is worth writing about here. 
</details>
OK. SGTM.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/robotlocomotion/drake/3276)*
<!-- Sent from Reviewable.io -->