@liangfok 2016-08-29
Review status: all files reviewed at latest revision, 1 unresolved discussion.

---

*[CHANGELOG.md, line 26 \[r3\]](https://reviewable.io:443/reviews/robotlocomotion/drake/3276#-KQJAilV7JEqslgqy08v:-KQJAilV7JEqslgqy08w:b-sihhoh) ([raw file](https://github.com/robotlocomotion/drake/blob/94a0f30d6cdaf313d326bcac44ec809e39b2a03b/CHANGELOG.md#L26)):*
> ```Markdown
> ### Changed
> 
>  - [#3276][] The `measure::execution` function has been replaced by `MeasureExecutionTime`.
> ```

BTW: nit: Added `()` behind method names to indicate that they are method names.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/robotlocomotion/drake/3276)*
<!-- Sent from Reviewable.io -->