@jwnimmer-tri 2016-09-06
+@david-german-tri for platform review, please.