@david-german-tri 2016-09-06
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Review status: all files reviewed at latest revision, 1 unresolved discussion.

---

*[CHANGELOG.md, line 39 \[r4\]](https://reviewable.io:443/reviews/robotlocomotion/drake/3276#-KR-G5EQX6PIbT-POMgO:-KR-G5EQX6PIbT-POMgP:b-o7ywv1) ([raw file](https://github.com/robotlocomotion/drake/blob/1b0a3f2f3e8768002f431ba728449455f2e321aa/CHANGELOG.md#L39)):*
> ```Markdown
>  - [#2984][] Renamed and moved `Polynomial.h` and `TrigPoly.h` from `drake/util` to `drake/common` and into the `drakeCommon` library.
>  - [#2963][] Rename RigidBody::CollisionElement to RigidBodyCollisionElement.
>  - [#3010][] All header file names under `solvers` are now spelled with lower case and underscore names.
> ```

why deleted?

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/robotlocomotion/drake/3276#-:-KR-Gl4D4wrUHF6Ox9OV:bnfp4nl)*
<!-- Sent from Reviewable.io -->